+++ src/main/java/com/naha/crimereportingsystem/admin/AdminController.java
@@ -5,12 +5,16 @@ import java.security.Principal;
 
 import com.naha.crimereportingsystem.citizens.Citizen;
 import com.naha.crimereportingsystem.citizens.CitizenService;
+import com.naha.crimereportingsystem.citizens.CitizenServiceImpl;
 import com.naha.crimereportingsystem.complaint.Complaint;
 import com.naha.crimereportingsystem.complaint.ComplaintService;
+import com.naha.crimereportingsystem.complaint.ComplaintServiceImpl;
 import com.naha.crimereportingsystem.police.Police;
 import com.naha.crimereportingsystem.police.PoliceService;
+import com.naha.crimereportingsystem.police.PoliceServiceImpl;
 import com.naha.crimereportingsystem.user.User;
 import com.naha.crimereportingsystem.user.UserService;
+import com.naha.crimereportingsystem.user.UserServiceImpl;
 
 import org.springframework.ui.Model;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,16 +30,16 @@ import org.springframework.web.bind.annotation.RequestParam;
 public class AdminController {
 
     @Autowired
-    UserService userService;
+    UserServiceImpl userService;
 
     @Autowired
-    PoliceService policeService;
+    PoliceServiceImpl policeService;
 
     @Autowired
-    CitizenService citizenService;
+    CitizenServiceImpl citizenService;
 
     @Autowired
-    ComplaintService complaintService;
+    ComplaintServiceImpl complaintService;
 
     @GetMapping("/admin")
     public String adminLoginRouter(Principal principal) {
