+++ src/main/java/com/naha/crimereportingsystem/complaint/ComplaintService.java
@@ -8,27 +8,13 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
-@Service
-public class ComplaintService {
+public interface ComplaintService {
 
-    @Autowired
-    ComplaintRepository complaintRepository;
+    public List<Complaint> findAllComplaintDetails();
 
-    public List<Complaint> findAllComplaintDetails() {
-        List<Complaint> complaints = new ArrayList<>();
-        complaintRepository.findAll().forEach(complaints::add);
-        return complaints;
-    }
+    public Complaint findComplaintDetailsById(long id);
 
-    public Complaint findComplaintDetailsById(long id) {
-        return complaintRepository.findById(id).orElse(null);
-    }
+    public Complaint saveComplaintDetails(Complaint complaint);
 
-    public Complaint saveComplaintDetails(Complaint complaint) {
-        return complaintRepository.save(complaint);
-    }
-
-    public void deleteComplaintDetail(long id) {
-        complaintRepository.deleteById(id);
-    }
+    public void deleteComplaintDetail(long id);
 }
