+++ src/main/java/com/naha/crimereportingsystem/crimes/CrimeService.java
@@ -6,24 +6,12 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-@Service
-public class CrimeService {
+public interface CrimeService {
 
-    @Autowired
-    CrimeRepository crimeRepository;
+    public void addCrimeDetail(Crime crime);
 
-    public void addCrimeDetail(Crime crime) {
-        crimeRepository.save(crime);
-    }
+    public void deleteCrimeDetail(long id);
 
-    public void deleteCrimeDetail(long id) {
-        crimeRepository.deleteById(id);
-    }
-
-    public List<Crime> findAllCrime() {
-        List<Crime> crimes = new ArrayList<>();
-        crimeRepository.findAll().forEach(crimes::add);
-        return crimes;
-    }
+    public List<Crime> findAllCrime();
 
 }
