+++ src/main/java/com/naha/crimereportingsystem/crimes/CrimeServiceImpl.java
@@ -1,0 +1,28 @@
+package com.naha.crimereportingsystem.crimes;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CrimeServiceImpl implements CrimeService {
+
+    @Autowired
+    CrimeRepository crimeRepository;
+
+    public void addCrimeDetail(Crime crime) {
+        crimeRepository.save(crime);
+    }
+
+    public void deleteCrimeDetail(long id) {
+        crimeRepository.deleteById(id);
+    }
+
+    public List<Crime> findAllCrime() {
+        List<Crime> crimes = new ArrayList<>();
+        crimeRepository.findAll().forEach(crimes::add);
+        return crimes;
+    }
+}
