+++ src/main/java/com/naha/crimereportingsystem/emergencyComplaint/EmergencyComplaintService.java
@@ -6,23 +6,11 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-@Service
-public class EmergencyComplaintService {
+public interface EmergencyComplaintService {
 
-    @Autowired
-    EmergencyComplaintRepository emergencyComplaintRepository;
+    public List<EmergencyComplaint> findAllEmergencyComplaintDetails();
 
-    public List<EmergencyComplaint> findAllEmergencyComplaintDetails() {
-        List<EmergencyComplaint> complaints = new ArrayList<>();
-        emergencyComplaintRepository.findAll().forEach(complaints::add);
-        return complaints;
-    }
+    public EmergencyComplaint findComplaintDetailsById(long id);
 
-    public EmergencyComplaint findComplaintDetailsById(long id) {
-        return emergencyComplaintRepository.findById(id).orElse(null);
-    }
-
-    public EmergencyComplaint saveComplaintDetails(EmergencyComplaint complaint) {
-        return emergencyComplaintRepository.save(complaint);
-    }
+    public EmergencyComplaint saveComplaintDetails(EmergencyComplaint complaint);
 }
