+++ src/main/java/com/naha/crimereportingsystem/home/HomeController.java
@@ -1,6 +1,7 @@
 package com.naha.crimereportingsystem.home;
 
 import com.naha.crimereportingsystem.crimes.CrimeService;
+import com.naha.crimereportingsystem.crimes.CrimeServiceImpl;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -11,7 +12,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 public class HomeController {
 
     @Autowired
-    CrimeService crimeService;
+    CrimeServiceImpl crimeService;
 
     @GetMapping("/")
     public String HomeRoute(Model model) {
